new holder class for schema update types
